Update to libX11 1.7.0. Tested by gkoehler@ and jsg@

+++ lib/libX11/src/WMProps.c
@@ -80,7 +80,7 @@ void XSetWMProperties (
      char **argv,		/* command line */
      int argc,			/* size of command line */
      XSizeHints *sizeHints,	/* size hints for window in its normal state */
-     XWMHints *wmHints,		/* miscelaneous window manager hints */
+     XWMHints *wmHints,		/* miscellaneous window manager hints */
      XClassHint *classHints)	/* resource name and class */
 {
     XTextProperty textprop;
@@ -105,7 +105,7 @@ void XSetWMProperties (
     textprop.value = (unsigned char *) hostName;
     textprop.encoding = XA_STRING;
     textprop.format = 8;
-    textprop.nitems = len;
+    textprop.nitems = (unsigned long) len;
     XSetWMClientMachine (dpy, w, &textprop);
 
     /* set hints about how geometry and window manager interaction */
@@ -142,6 +142,6 @@ void XSetWMProperties (
     if (locale)
         XChangeProperty (dpy, w, XInternAtom(dpy, "WM_LOCALE_NAME", False),
         XA_STRING, 8, PropModeReplace,
-        (unsigned char *)locale, strlen(locale));
+        (unsigned char *)locale, (int) strlen(locale));
 }
